New Diffusion class and methods
The Diffusion class contains methods coinFlipper() which flips a coin N times and gives the final position of an object that moves right given coin = 0 and left given coin = 1. The display method displays the trial number, number of steps taken, and final position of the object for each trial. The input method contains the inputs for variables that are more conveniently inputted in the terminal.